LayoutTests/imported/w3c/web-platform-tests/streams/transferable/window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_array_equals: there should be no ports in the event lengths differ, expected array [] length 0, got [object "[object MessagePort]"] length 1
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_array_equals: there should be no ports in the event lengths differ, expected array [] length 0, got [object "[object MessagePort]"] length 1

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: there should be no ports in the event lengths differ, expected array [] length 0, got [object "[object MessagePort]"] length 1

FAIL window.postMessage should be able to transfer a ReadableStream The object can not be cloned.
FAIL port.postMessage should be able to transfer a ReadableStream The object can not be cloned.
FAIL the same ReadableStream posted multiple times should arrive together The object can not be cloned.
FAIL transfer to and from an iframe should work promise_test: Unhandled rejection with value: object "DataCloneError: The object can not be cloned."
